Override version number in banner for `develop` builds on Docker.
This updates the Dockerfile to take in and maintain the SOURCE_COMMIT
and DOCKER_TAG environment variables, and updates banner.js to look for
those environment variables first determining the version. This results
in the version number being replaced with `develop-abcdefgh`, where
`abcdefgh` are the first 8 characters of the commit hash.