Rework internationalization of numbers.
This has some implications in how things are printed, so a few tests
were changed and a new set of lines added to weight-remove.test.log to
either make the tests match the new code or to test some formatting that
was not being tested.

Specifically, this removes the ">99.9%" and "<0.01%" output in favor of
simply rounding through i18n options. This is a relatively significant
change to the bot's output, but should be more precise and (at least in
my opinion) more readable.