Update Twurple->7.0.0.
This is effectively an alertnate method to address #115, which keeps the
same logic we were using before that PR. As such, that PR was reverted
prior to implementing this upgrade.

Additionally, this updates some dev dependencies that themselves depend
on `semver` to fix a vulnerability there. It doesn't impact us, but it
needs to be updated anyway.